+++ src/lib/browser/index.ts
@@ -1,0 +1,6 @@
+export {
+	BeforeUrlChangeEvent,
+	UrlChangeEvent,
+	eventBeforeUrlChange,
+	eventURLChange
+} from './navigationEvents/Events';