+++ src/lib/native/createHistory/HistoryType.ts
@@ -1,0 +1,6 @@
+import { type NavigationActionType } from '../../types/navigationActionType/NavigationActionType.native';
+
+export type NavigationState = {
+	action: NavigationActionType
+	url: URL
+};