+++ src/lib/types/navigationActionType/NavigationActionType.ts
@@ -1,0 +1,3 @@
+import { type BaseNavigationActionType } from './BaseNavigationActionType';
+
+export type NavigationActionType = BaseNavigationActionType;