+++ src/lib/utils/constants.ts
@@ -1,3 +1,0 @@
-export const IS_BROWSER: boolean = typeof window !== 'undefined' && typeof window.document !== 'undefined';
-
-export const ORIGIN = IS_BROWSER ? window.location.origin : 'http://localhost';