+++ src/test/App.test.tsx
@@ -1,6 +1,7 @@
+
 it('App', () => {
 	expect(true).toBeTruthy();
-});
+
 
 /*
 const object1 = {